syngfn-mfa/requirements.txt
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.13.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.12.0
bleach==5.0.1
certifi==2022.12.7
click==8.1.3
cmake==3.27.2
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.16.2
fastprogress==1.0.3
filelock==3.12.2
flake8==6.0.0
fqdn==1.5.1
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.30
hpack==4.0.0
hydra-core==1.3.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipdb==0.13.11
ipykernel==6.20.1
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.11.4
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
lazy-object-proxy==1.9.0
lit==16.0.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
multipledispatch==0.6.0
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
overrides==7.4.0
packaging==23.0
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
plotly==5.12.0
polyleven==0.8
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.21.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.14.0
pylint==2.15.10
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
rdkit==2023.3.3
referencing==0.30.2
reportlab==3.5.68
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
scikit-learn==1.2.0
scipy==1.10.0
selfies==2.1.1
Send2Trash==1.8.0
sentry-sdk==1.13.0
setproctitle==1.3.2
smart-open==6.4.0
SmilesPE==0.0.3
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tenacity==8.1.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
torch==1.12.1+cu113
torch-scatter==2.1.2
torch-sparse==0.6.18
torch_geometric==2.4.0
torchaudio==0.12.1+cu113
torchtyping==0.1.4
torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
triton==2.0.0
uri-template==1.2.0
urllib3==1.26.14
vina==1.2.5
wandb==0.17.4
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
wrapt==1.14.1
xtb==22.1
zipp==3.11.0
zstandard==0.22.0
